refactor(pragma-cli): keep three cheap seams off their domain barrels
Three reaches this branch rerouted put heavy modules on paths that never
needed them. The laziness filter I applied mirrors `lazy.test.ts` and
`safety.test.ts` exactly — so it catches what those guards cover and nothing
else, and none of these three is covered.

The hazard is the same each time and it is second-order: a barrel re-exports
its whole domain, so routing ONE cheap leaf import through it drags the
domain's expensive siblings onto the static graph.

`sources status` reads two manifests and prints them. Through
`graphpack/index.js` it would have initialised the entire pack stack — the
barrel re-exports `build.ts`/`read.ts`, pulling `@canonical/ke` and
`@canonical/ke-graphql`, and `embedded.ts` with its ~1.9 MB generated payload.

`sources update` builds a USER pack, and would have loaded the embedded one to
do it: `graphpack/index.js` re-exports `embedded.ts`, and with per-file `tsc`
output ESM evaluates that re-export when the module loads.

`setupConfig` computes a filename. Through `config/index.js` that reaches
`readConfig` → `defaults.ts` → `schema.ts` → `zod`, and this module sits on the
setup/doctor target table, so both commands would have paid for the config
parser to learn a path.

All three are back to direct leaf imports, each carrying the reason at the
import so a later sweep does not "fix" it again. No guard was relaxed and no
budget raised: perf 22/22, suite 1463 passed.
